model updated
seems that price changed from int to double, also updated android dependencies, but not necesary

+++ app/src/main/java/iesharia/retrofit2example/data/db/Product.kt
@@ -10,7 +10,7 @@ data class Product(
     var id: Int,
     var title: String,
     var description: String,
-    var price: Int,
+    var price: Double,
     var discountPercentage: Double,
     var rating: Double,
     var stock: Int,
