+++ app/src/main/java/iesharia/retrofit2example/data/network/ProductResponse.kt
@@ -7,7 +7,7 @@ data class ProductResponse(
     @SerializedName("id") var id: Int,
     @SerializedName("title") var title: String,
     @SerializedName("description") var description: String,
-    @SerializedName("price") var price: Int,
+    @SerializedName("price") var price: Double,
     @SerializedName("discountPercentage") var discountPercentage: Double,
     @SerializedName("rating") var rating: Double,
     @SerializedName("stock") var stock: Int,
